upgrade: remove unnecessary dependencies (#600)
* Inline color translation library

* Replace Epoch.net with standard features

* Remove unused Yoh.Text.Json.NamingPolicies dependency

* Revert "Inline color translation library"

This reverts commit 66b32aaebb417cff719a65cc38b3d2227bc5a051.

+++ Vonage/Common/ITimeProvider.cs
@@ -10,7 +10,7 @@ public interface ITimeProvider
     /// <summary>
     /// Returns the current epoch value.
     /// </summary>
-    int Epoch { get; }
+    long Epoch { get; }
 
     /// <summary>
     /// Returns the current date and time.
